requirements.txt
asgiref==3.3.4
Django==3.1.6
django-appconf==1.0.4
django-phonenumber-field==5.2.0
django-phonenumbers==1.0.1
django-tinymce4-lite==1.7.5
jsmin==2.2.2
phonenumbers==8.12.25
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
